# Make Notifications::Fanout faster and safer